tackle/providers/yubikey/requirements.txt
python-yubico>=1.3.0
